src/requirements.txt
aniso8601==6.0.0
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
chromedriver-binary==76.0.3809.68.0
Click==7.0
decorator==4.4.0
Flask==1.0.2
flask-restplus==0.12.1
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-pubsub==0.45.0
google-cloud-storage==1.15.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.12.3
grpcio==1.23.0
idna==2.8
ImageHash==4.0
ipython==7.7.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
numpy==1.16.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
Pygments==2.4.2
pyrsistent==0.15.2
python-magic==0.4.15
pytz==2019.1
PyWavelets==1.0.3
requests==2.21.0
rsa==4.0
scipy==1.2.1
selenium==3.141.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
Werkzeug==0.15.4
